First commit.
This program was originally written in the early years of 2006.  It was ported from a C implementation, which was not configurable, but utilized the same stemming algorithm (Porter) used to normalize plurality in common words.  Additionally, the C implementation made use of a multimap data structure, still missing in Java.

The chosen and adhered to approach was to maintain the original C implementation during port to Java.  The Porter Java implementation was adopted, and not touched and/or modified.  In addition, stop words and stop characters were maintained.

New features from the original implementation include: full input and output ancillary configuration; a variety of command line options; export not only to tab-delimited text, but also to Excel (where Retired Concept result(s) are bolded in the color Red); the addition of the owl-api for OWL1; ad-hoc String parsing of OWL1 "complex properties" (i.e., rdf:ParseType Literals and datatypes XMLLiterals); a Perl script to generate an OWL